client: Extract colorchooser into a separate component
